+++ frontend/src/views/index/AccountSettings.vue
@@ -7,8 +7,8 @@ import { useGlobalState } from '../../store'
 import { api } from '../../api'
 
 const {
-    jwt, localeCache, settings, showPassword, loading,
-    mailboxSplitSize, useIframeShowMail
+    jwt, localeCache, settings, showAddressCredential, loading,
+    mailboxSplitSize, useIframeShowMail, preferShowTextMail
 } = useGlobalState()
 const router = useRouter()
 const message = useMessage()
@@ -20,20 +20,22 @@ const { t } = useI18n({
     messages: {
         en: {
             mailboxSplitSize: 'Mailbox Split Size',
-            useIframeShowMail: 'Use iframe Show Mail',
+            useIframeShowMail: 'Use iframe Show HTML Mail',
+            preferShowTextMail: 'Display text Mail by default',
             logout: "Logout",
             delteAccount: "Delete Account",
-            showPassword: 'Show Password',
+            showAddressCredential: 'Show Address Credential',
             logoutConfirm: 'Are you sure to logout?',
             delteAccount: "Delete Account",
             delteAccountConfirm: "Are you sure to delete your account and all emails for this account?",
         },
         zh: {
             mailboxSplitSize: '邮箱界面分栏大小',
-            useIframeShowMail: '使用iframe显示邮件',
+            preferShowTextMail: '默认以文本显示邮件',
+            useIframeShowMail: '使用iframe显示HTML邮件',
             logout: '退出登录',
             delteAccount: "删除账户",
-            showPassword: '查看密码',
+            showAddressCredential: '查看邮箱地址凭证',
             logoutConfirm: '确定要退出登录吗？',
             delteAccount: "删除账户",
             delteAccountConfirm: "确定要删除你的账户和其中的所有邮件吗?",
@@ -72,12 +74,15 @@ const deleteAccount = async () => {
                         0.75: '0.75'
                     }" />
                 </n-form-item-row>
+                <n-form-item-row :label="t('preferShowTextMail')">
+                    <n-switch v-model:value="preferShowTextMail" :round="false" />
+                </n-form-item-row>
                 <n-form-item-row :label="t('useIframeShowMail')">
                     <n-switch v-model:value="useIframeShowMail" :round="false" />
                 </n-form-item-row>
             </n-card>
-            <n-button @click="showPassword = true" type="primary" secondary block strong>
-                {{ t('showPassword') }}
+            <n-button @click="showAddressCredential = true" type="primary" secondary block strong>
+                {{ t('showAddressCredential') }}
             </n-button>
             <n-button @click="showLogout = true" secondary block strong>
                 {{ t('logout') }}