+++ worker/src/constants.js
@@ -2,4 +2,5 @@ export const CONSTANTS = {
     ADDRESS_BLOCK_LIST_KEY: 'address_block_list',
     SEND_BLOCK_LIST_KEY: 'send_block_list',
     AUTO_CLEANUP_KEY: 'auto_cleanup',
+    USER_SETTINGS_KEY: 'user_settings',
 }