+++ worker/src/scheduled.js
@@ -1,15 +1,16 @@
 import { cleanup } from './common'
 import { CONSTANTS } from './constants'
 import { getJsonSetting } from './utils';
+import { CleanupSettings } from './models';
 
 export async function scheduled(event, env, ctx) {
     console.log("Scheduled event: ", event);
-    let autoCleanupSetting = await getJsonSetting(
+    const value = await getJsonSetting(
         { env: env, },
         CONSTANTS.AUTO_CLEANUP_KEY
     );
+    const autoCleanupSetting = new CleanupSettings(value);
     console.log("autoCleanupSetting:", JSON.stringify(autoCleanupSetting));
-    autoCleanupSetting = autoCleanupSetting || {};
     if (autoCleanupSetting.enableMailsAutoCleanup && autoCleanupSetting.cleanMailsDays > 0) {
         await cleanup(
             { env: env, },